% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{Add a gitignore file}
\usage{
add_gitignore(path = NULL)
}
\arguments{
\item{path}{optionally the path to a directory, if not set then
the current directory/RStudio project is set as the destination}
}
\description{
Add a gitignore file
}
