% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_quarto_ignores}
\alias{add_quarto_ignores}
\title{Add quarto ignores to a gitignore file}
\usage{
add_quarto_ignores(gitignore = NULL)
}
\arguments{
\item{gitignore}{optionally the path to a gitignore file, if not set
then the gitignore in the current directory/RStudio project is located}
}
\description{
Add quarto ignores to a gitignore file
}
