% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{checkUserCredentials}
\alias{checkUserCredentials}
\title{checkPassword}
\usage{
checkUserCredentials(con, username, password)
}
\arguments{
\item{con}{Database connection}

\item{username}{Username}

\item{password}{Password for the user}
}
\description{
Check whether the password for the user matches the hashed (using scrypt)
password in the database.
}
\examples{
\dontrun{
checkUserCredentials(con, "foo", "42")
}
}
