% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_global.R
\name{clear_global}
\alias{clear_global}
\title{Clear the global environment}
\usage{
clear_global()
}
\description{
A function to quickly clear the global environment from the
console.
}
\examples{
clear_global()
}
