% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_all.R
\name{close_all}
\alias{close_all}
\title{Close all open source files in RStudio}
\usage{
close_all()
}
\value{
invisibly the number of windows closed
}
\description{
This isn't necessary as is covered by an in-built function
within the RStudio IDE and can be called by `cmd`+`shift`+`w` on Mac.
}
