% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{configure}
\alias{configure}
\title{configure}
\usage{
configure()
}
\value{
the configuration
}
\description{
Load the default configuration and the configuration defined in the
\code{config} file in the working directory.
}
\details{
The default configuration is retrieved from \code{defaults/config.R} in the
package. Then settings found in the variable named \code{config} in
\code{config.R} in the working directory is used to overwrite to optionally
overwrite settings.
}
\examples{
\dontrun{
configure()
}
}
