% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sf.R
\name{count_sf}
\alias{count_sf}
\title{Quicker counting of sf objects}
\usage{
count_sf(x, ...)
}
\arguments{
\item{x}{An `sf` object}

\item{...}{Arguments to pass on to `dplyr::count()`}
}
\description{
[dplyr::count()] can take a long time on an `sf` object and may well fail
due to collation problems. The `sf` package provides a method for `count()`
which adds a `.drop_geometry` argument to reduce the run time. `count_sf()`
is a more explicit alternative that will always drop the geometry.
}
\details{
For the `sf` method to work the package has to already be part of the call
stack (which, to be fair, is likely if working with `sf` objects). However,
you explicitly set `.drop_geometry = "TRUE"`, otherwise the `sf` method for
`dplyr::summarise()` will kick in which seeks to collate geometries for each
counting group, this creates delay and can fail if there are issues in the
collation process. `count_sf()` is intentionally designed to drop geometry
and only count features.
}
