% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{createRoute}
\alias{createRoute}
\title{Return a function that dispatches to `handler()` if `path` matches.}
\usage{
createRoute(path, handler)
}
\arguments{
\item{path}{URL path.}

\item{handler}{Handler for this request given the path.}
}
\value{
A response if the path is matched, NULL if the path is not matched.
}
\description{
Return a function that dispatches to `handler()` if `path` matches.
}
