% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{createRoutes}
\alias{createRoutes}
\title{createRoutes}
\usage{
createRoutes(...)
}
\arguments{
\item{...}{A list of c(url, handler)'s}
}
\description{
Create routes from a list of urls and handlers
}
\examples{
\dontrun{
createRoutes(
   c("index.html", function(...) 1),
   c("contact.html", function(...) 2)
}
}
