% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{createUser}
\alias{createUser}
\title{createUser}
\usage{
createUser(con, username, password)
}
\arguments{
\item{con}{Database connection}

\item{username}{Username}

\item{password}{Password for the user}
}
\description{
Create a user in the database. The password is hashed using scrypt.
}
\examples{
\dontrun{
createUser(con, "foo", "42")
}
}
