% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{errorResponse}
\alias{errorResponse}
\title{errorResponse}
\usage{
errorResponse(message = "")
}
\arguments{
\item{An}{optional message to show to the user.}
}
\description{
Response for requests that generate an error.
}
