% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{existsUser}
\alias{existsUser}
\title{existsUser}
\usage{
existsUser(con, username)
}
\arguments{
\item{con}{Database connection}

\item{username}{Username}
}
\description{
Check whether a users exists
}
\examples{
\dontrun{
createUser(con, "foo", "42")
}
}
