% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{extractParameters}
\alias{extractParameters}
\title{Extract Parameters from a request}
\usage{
extractParameters(request)
}
\arguments{
\item{request}{The request object from which to extract the parameters.}
}
\value{
A list of extracted parameters.
}
\description{
Extract the parameters from a request (GET, POST and from the URL).
}
