% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileUrl}
\alias{fileUrl}
\title{fileUrl
Return a URL for a file to be sent to the client. the file will be base64
encoded and embedded in the URL.}
\usage{
fileUrl(file, contentType = "application/octet-stream")
}
\arguments{
\item{file}{The file name}

\item{contentType}{The content type}
}
\description{
Taken from
https://github.com/rstudio/shiny/blob/e30fac02eddbb7cd510d448e385c559c82009a24/R/shiny.R
}
