% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{genericView}
\alias{genericView}
\title{Respond with the result of FUN.}
\usage{
genericView(FUN)
}
\arguments{
\item{FUN}{Function that generates the response as text.}
}
\value{
A response object.
}
\description{
Respond with the result of FUN.
}
\examples{
genericView(function() "Hello World!")
}
