% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{getConfigOrDefault}
\alias{getConfigOrDefault}
\title{Get a value from the config or return the default}
\usage{
getConfigOrDefault(config, key, default)
}
\arguments{
\item{config}{The configuration list}

\item{key}{The key to look for}

\item{default}{The default}
}
\value{
The key if found or the default
}
\description{
Get a value from the config or return the default
}
\examples{
getConfigOrDefault(configure, port, 8888)
}

