% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{getRandomString}
\alias{getRandomString}
\title{getRandomString}
\usage{
getRandomString(length)
}
\arguments{
\item{length}{The length of the random output.}
}
\value{
A base64 encoded number with \code{length} digits.
}
\description{
Return a base64 encoded random number of the given lenght. The length must be
a multiple of 4.
}
\details{
The source of randomness is \code{openssl::rand_byes}.
}
\examples{
getRandomString(16)
}
