% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{getResponse}
\alias{getResponse}
\title{Get a response}
\usage{
getResponse(resp, req)
}
\arguments{
\item{resp}{The response object moving through the middlewares.}

\item{req}{The request for which the response is being created.}
}
\description{
Handle the middlewares, as is done in Django.
}
