% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{getSessionData}
\alias{getSessionData}
\title{getSessionData}
\usage{
getSessionData(con, sessionKey)
}
\arguments{
\item{con}{Connection to database.}

\item{sessionKey}{The session key for which to get session data.}
}
\value{
The session data.
}
\description{
getSessionData
}
\examples{
\dontrun{
getSessionData(con, "123")
}
}
