% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{helpMessage}
\alias{helpMessage}
\title{Display the help message}
\usage{
helpMessage(...)
}
\arguments{
\item{...}{Ignored, just show the default help message.}
}
\description{
Display the help message
}
\examples{
\dontrun{
helpMessage()
}
}
