% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihme_cols.R
\docType{data}
\name{ihme_cols}
\alias{ihme_cols}
\title{Vector of colors that I like}
\format{An object of class \code{character} of length 10.}
\usage{
ihme_cols
}
\description{
Taken from an IHME visualization
}
\keyword{datasets}
