% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{isMattRRunning}
\alias{isMattRRunning}
\title{isMattRRunning}
\usage{
isMattRRunning()
}
\value{
Whether the server is running.
}
\description{
Check whether a daemonized server is running.
}
\examples{
isMattRRunning()
}
