% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{killByToken}
\alias{killByToken}
\title{Kill a process by the given token.}
\usage{
killByToken(token)
}
\arguments{
\item{token}{Token of the process to kill.}
}
\value{
TRUE on success, FALSE on failure.
}
\description{
Kill a process by the given token.
}

