% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{mattR_admin}
\alias{mattR_admin}
\title{mattR_admin}
\usage{
mattR_admin(routes, appState)
}
\arguments{
\item{routes}{The original set of routes.}

\item{appState}{The state of the app.}
}
\description{
Provide an admin interface.
}
