% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Safe merge: check column intersection before attempting merge}
\usage{
merge(x, y, by = intersect(names(x), names(y)), by.x = by, by.y = by,
  all = FALSE, all.x = all, all.y = all, sort = TRUE,
  suffixes = c(".x", ".y"), no.dups = TRUE, incomparables = NULL,
  check = TRUE, ...)
}
\arguments{
\item{check}{should a check be performed?}
}
\value{
Either merged dataframes, or NULL if aborted
}
\description{
Keeps all params as base::merge()
}
