% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_ct.R
\name{na_ct}
\alias{na_ct}
\title{Get count of NAs in each column in a dataframe or tibble}
\usage{
na_ct(df)
}
\arguments{
\item{df}{data frame}
}
\value{
named vector of NA counts
}
\description{
Get count of NAs in each column in a dataframe or tibble
}
