% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{notFoundResponse}
\alias{notFoundResponse}
\title{notFoundResponse}
\usage{
notFoundResponse(message = "")
}
\arguments{
\item{message}{An optional message to show to the user.}
}
\description{
Response for objects (URL's) that can't be found.
}
