% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{parseCookie}
\alias{parseCookie}
\title{parseCookie}
\usage{
parseCookie(http_cookie)
}
\arguments{
\item{http_cookie}{The cookie as a string.}
}
\value{
The cookie parsed to a list. The empty list is returned when an error
or warning occurs and a \code{message()} is generated.
}
\description{
Parse a HTTP cookie
}
\examples{
parseCookie("foo=bar")
}
