% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/+.R
\name{+}
\alias{+}
\title{Redefine + operator to concatenate strings, like in python}
\usage{
"+"()
}
\arguments{
\item{two}{variables}
}
\value{
either the some of two numeric or the concatenation of two character
}
\description{
Redefine + operator to concatenate strings, like in python
}
