% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragg_png.R
\name{ragg_png}
\alias{ragg_png}
\title{Easy ragg-based plots}
\usage{
ragg_png(file, plot = NULL, width = 600, height = 600, overwrite = NULL, ...)
}
\arguments{
\item{file}{The file name of the image}

\item{plot}{The plot to save, uses `ggplot2::last_plot()` if left as `NULL`}

\item{width}{The width in pixels}

\item{height}{The height in pixels}

\item{overwrite}{Whether to overwrite existing files}

\item{...}{Arguments passed down to `ragg::agg_png()`}
}
\description{
A [ggplot2::ggsave] equivalent using the `ragg` package to produce
better png images.
}
