% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{readPartials}
\alias{readPartials}
\title{Read partial templates for template rendering}
\usage{
readPartials(directory)
}
\arguments{
\item{directory}{Directory of the partial templates}
}
\value{
A named list of partials.
}
\description{
Read partial templates for template rendering
}
