% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{runTestServer}
\alias{runTestServer}
\title{Run a server for testing mattR apps}
\usage{
runTestServer(daemonized = FALSE)
}
\arguments{
\item{daemonized}{Whether to start the server daemonized.}
}
\value{
The httpuv handle to the server process.
}
\description{
Run a server for testing mattR apps
}
\examples{
\dontrun{
runTestServer()
}
}
