% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{sessionMiddleware}
\alias{sessionMiddleware}
\title{sessionMiddleware}
\usage{
sessionMiddleware(resp, req)
}
\arguments{
\item{resp}{The response that is being build.}

\item{req}{The request.}
}
\description{
Read the cookie and create a session id.
}
