% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{setSessionData}
\alias{setSessionData}
\title{setSessionData}
\usage{
setSessionData(con, sessionKey, sessionData)
}
\arguments{
\item{con}{Connection to database.}

\item{sessionKey}{Session key to set session data for.}

\item{sessionData}{The session data to set.}
}
\value{
The session data.
}
\description{
setSessionData
}
\examples{
\dontrun{
setSessionData(con, "123", "foo=bar")
}
}
