% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{setupDatabase}
\alias{setupDatabase}
\title{setupDatabase}
\usage{
setupDatabase()
}
\value{
A connection to the database
}
\description{
Setup a database with a simple USERS table.
}
\examples{
\dontrun{
setupDatabase()
}
}
