% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{skeleton}
\alias{skeleton}
\title{Create a skeleton for the app}
\usage{
skeleton(path = getwd())
}
\arguments{
\item{path}{Target path for skeleton creation}
}
\value{
\code{TRUE} on success, \code{FALSE} otherwise.
}
\description{
Create a skeleton for the app
}
\examples{
\dontrun{
buildApp()
}
}
