% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{startServerProcess}
\alias{startServerProcess}
\title{Start a server as an external process, killable by killByToken(output)}
\usage{
startServerProcess()
}
\value{
The probably unique token of the process.
}
\description{
Start a server as an external process, killable by killByToken(output)
}
\examples{
\dontrun{
startServerProcess()
}
}

