% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{startTestServer}
\alias{startTestServer}
\title{startTestServer}
\usage{
startTestServer(app, host = "0.0.0.0", port = 8080L, daemonized = FALSE,
  debug = FALSE)
}
\arguments{
\item{app}{A rook app.}

\item{host}{The host to bind on.}

\item{port}{The port to listen on.}

\item{daemonized}{Whether to start the server daemonized.}

\item{debug}{Whether debug is activated.}
}
\description{
Run a test server based on the given settings.
}
