% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{staticView}
\alias{staticView}
\title{Return static files}
\usage{
staticView(staticDir, urlPath)
}
\arguments{
\item{staticDir}{Location of the static files on the server}

\item{urlPath}{URL path to the static files}
}
\value{
A response object or NULL if the file is not found.
}
\description{
Return static files
}
\examples{
\dontrun{
staticFun <- staticView("/var/www/static/", "/static/")
}
}

