% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{stopDaemonizedServer}
\alias{stopDaemonizedServer}
\title{stopDaemonizedServer}
\usage{
stopDaemonizedServer()
}
\value{
\code{invisible()}
}
\description{
Stop a daemonized server
}
\examples{
\dontrun{
stopDaemonizedServer()
}
}
