% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{storeSessionId}
\alias{storeSessionId}
\title{Store a session id to the database}
\usage{
storeSessionId(con, sessionKey)
}
\arguments{
\item{con}{A connection to the database.}

\item{sessionKey}{The session key for current request.}
}
\value{
The result of the DB query saving to the database.
}
\description{
Store a session id to the database
}
