% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Like table() but with useNA set to 'ifany' by default}
\usage{
table(...)
}
\arguments{
\item{...}{vector}
}
\value{
table object
}
\description{
Like table() but with useNA set to 'ifany' by default
}
