% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{templateView}
\alias{templateView}
\title{Respond with a rendered template}
\usage{
templateView(templateFile, data)
}
\arguments{
\item{templateFile}{Filepath of a whisker template}

\item{data}{named list or environment with variables that will be used during
rendering}
}
\value{
A response object with as body the rendered template.
}
\description{
Respond with a rendered template
}
\examples{
templateView(system.file("static", "index.html", package = "mattR"),
             list(title = "foo", text = "bar"))
}
