% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{buildApp}
\alias{buildApp}
\title{Build a mattR app object for serving by httpuv}
\usage{
buildApp(config)
}
\value{
The app object
}
\description{
Build a mattR app object for serving by httpuv
}
\examples{
\dontrun{
buildApp()
}
}

