% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{buildApp}
\alias{buildApp}
\title{buildApp}
\usage{
buildApp(routes, modules = c(), appState = new.env())
}
\arguments{
\item{routes}{The routes used in this app.}

\item{modules}{List of optional modules.}

\item{appState}{The initial state of the app.}

\item{config}{List containing the configuration for this app.}
}
\value{
The app object
}
\description{
Build a mattR app object for serving by httpuv
}
\details{
This will set the variable in the appState environment if it is not set.
}
\examples{
\dontrun{
buildApp()
}
}
