% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{createConnection}
\alias{createConnection}
\title{createConnection}
\usage{
createConnection()
}
\value{
Connection to the database.
}
\description{
Create a connection to an in-memory SQLite database.
}
\examples{
\dontrun{
createConnection()
}
}
