% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{guessContentTypeFromFilePath}
\alias{guessContentTypeFromFilePath}
\title{Guess the Content-Type for a file or filepath}
\usage{
guessContentTypeFromFilePath(filepath)
}
\arguments{
\item{filepath}{Filepath of which to guess the Content-Type.}
}
\value{
The guess of the Content-Type
}
\description{
Guess the Content-Type for a file or filepath
}
