% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matt_prompt.R
\name{matt_prompt}
\alias{matt_prompt}
\title{Customised R prompt}
\usage{
matt_prompt(...)
}
\description{
A function for use in an .Rprofile to update the R console prompt.
}
\details{
Add the following to the .First() function in your .Rprofile file:
  mattR::matt_prompt()
  addTaskCallback(mattR::matt_prompt)
}
