% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{renderTemplate}
\alias{renderTemplate}
\title{Render a template given data}
\usage{
renderTemplate(templateFile, data)
}
\arguments{
\item{templateFile}{Filepath of a whisker template.}

\item{data}{Named list or environment with variables that will be used during
rendering.}
}
\value{
The rendered template.
}
\description{
Render a template given data
}
\examples{
renderTemplate(system.file("static", "index.html", package = "mattR"),
               list(title = "foo", text = "bar"))
}
