% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{setupResponse}
\alias{setupResponse}
\title{Set up response}
\usage{
setupResponse()
}
\value{
An empty request with middlewares attached.
}
\description{
Set up an empty response to be handled by the attached middlewares.
}
