% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{startTestServer}
\alias{startTestServer}
\title{startTestServer}
\usage{
startTestServer(app, appState = new.env(), host = "0.0.0.0",
  port = 8080L, daemonized = FALSE)
}
\arguments{
\item{app}{A rook app.}

\item{appState}{State of the app.}

\item{host}{The host to bind on.}

\item{port}{The port to listen on.}

\item{daemonized}{Whether to start the server daemonized.}
}
\description{
Run a test server based on the given settings.
}
