% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sf.R
\name{tidy_sf}
\alias{tidy_sf}
\title{Coerce sf data frames into tibbles}
\usage{
tidy_sf(x, .verbose = FALSE)
}
\arguments{
\item{x}{An `sf::st_sf()` object}

\item{.verbose}{Whether to be noisy in checks}
}
\description{
A [tibble::tibble()] prints more nicely in the console than a `data.frame()`.
`sf`` objects can often contain many rows, making printing to the console
cumbersome. [sf::read_sf()] reads in geospatial data as tibbles,
[sf::st_read()] produces a base data.frame. `sf` objects in packages or from
other sources might not be tibble-like, `tidy_sf()` provides a simple
utility for use in pipelines to easily convert an `sf` object to a tibble
format.
}
